Extract Storm Setup detail presentation builders #301
